fix: Better Secure Validation for analysis and delete

+++ src/main/java/com/shortener/api/ShortenerApi.java
@@ -3,6 +3,7 @@ package com.shortener.api;
 import com.shortener.dto.CreateUrlRequest;
 import com.shortener.dto.ShortenResponse;
 import com.shortener.dto.UrlStats;
+import com.shortener.errors.ShortIdNotFound;
 import com.shortener.model.Url;
 import com.shortener.service.UrlService;
 import jakarta.validation.Valid;
@@ -12,6 +13,8 @@ import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+
+import java.util.regex.Pattern;
 
 
 @Slf4j
@@ -24,6 +27,10 @@ public class ShortenerApi {
 
     @Value("${app.base-url}")
     private String baseUrl;
+
+    final Pattern PATTERN =
+            Pattern.compile ("^[A-Za-z0-9_-]{3,10}$");
+
 
     @GetMapping("/ping")
     public String ping() {
@@ -46,8 +53,13 @@ public class ShortenerApi {
         return ResponseEntity.status(HttpStatus.CREATED).body(response);
     }
 
+
     @GetMapping("/{shortId}/analysis")
     public ResponseEntity<UrlStats> analysis(@PathVariable("shortId") String shortId){
+
+        if (!PATTERN.matcher(shortId).matches()) {
+                throw new ShortIdNotFound("Invalid shortId");
+        }
         UrlStats urlStats = urlService.getStats(shortId);
         urlStats.setShortUrl(baseUrl+ "/" + shortId);
         return ResponseEntity.status(HttpStatus.OK).body(urlStats);
@@ -55,6 +67,11 @@ public class ShortenerApi {
 
     @DeleteMapping("/{shortId}")
     public ResponseEntity<String> delete(@PathVariable("shortId") String shortId){
+
+        if (!PATTERN.matcher(shortId).matches()) {
+            throw new ShortIdNotFound("Invalid shortId");
+        }
+
         String response = urlService.delete(shortId);
         return ResponseEntity.status(HttpStatus.NO_CONTENT)
                 .body(response+ " is deleted");
